Shark chase: one copy of the bots
`bot-strategy.ts` was written twice, the copies sharing about 180 of
their 210 lines: `asSharkRoute`, the whole random bot, the memoised
survival search and the location preference differed only in `16`/`25`
and `11`/`15`.

`makeSharkBots` in the new `bot-search.ts` takes both bots, as
`makeGeometry(size)` and `makeGameplay(maxTurn)` already take what they
need. What is passed in is what is genuinely per-variant: the size of
the lake, the last day, the sectors the shark prefers when nothing it
can reach is safe, the researchers' scripted line, and — on 5 × 5 — the
precomputed answer that spares the live search on the early days it is
too slow for. Each call keeps its own memo, as each module did.

This is movement, not design. The preference rings stay exactly the two
hand-written lists, now passed in rather than inlined; both lakes are
played by the same moves, so `Moves` becomes one type, which is what
lets the two bots share a signature.

The one thing that is not movement: `getOptimalSubmarineMoveByBot`
opened with a capture check, and lifting the table out as it stood would
have had the bot follow its script straight past a submarine already
beside the shark. The capture happens in the shared `getSubmarineMove`
now, ahead of the script — the same order as before.
